requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1; python_version < '3.9'
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cx-oracle==8.3.0
defusedxml==0.7.1
dj-database-url==0.5.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-templated-mail==1.1.1
django==4.0.3
djangorestframework-simplejwt==4.8.0
djangorestframework==3.13.1
djoser==2.1.0
et-xmlfile==1.1.0
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-secret-manager==2.9.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.3
grpcio-status==1.44.0
grpcio==1.44.0
gunicorn==20.1.0
idna==3.3; python_version >= '3'
itypes==1.2.0
jinja2==3.0.3
markupsafe==2.1.1
numpy==1.22.3; platform_machine != 'aarch64' and platform_machine != 'arm64' and python_version < '3.10'
oauthlib==3.2.0
openpyxl==3.0.9
pandas==1.4.1
proto-plus==1.20.3
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pyjwt==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1
requests==2.27.1
rsa==4.8; python_version >= '3.6'
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
